Reintegrated tech branch (r2352 - r2958)
Mostly comment changes and some fixed warnings. 
All comments in guess.h converted to doxygen style which may give some conflicts,
but they should be simple to solve since it's only comments.

Bugfix in plib (some broken ins files could cause LPJ-GUESS to go into infinite memory eating loop)

benchmarks: Identical results.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/trunk@2959 70805675-57b6-444b-bb00-18ecffcb4be8

+++ framework/guessserializer.cpp
@@ -130,7 +130,7 @@ void verify_meta_data(const char* directory) {
 	// Verify that the PFTs have the same names
 	for (int i = 0; i < npft; i++) {
 
-		const int PFT_NAME_MAX_SIZE = 256;
+		const size_t PFT_NAME_MAX_SIZE = 256;
 
 		unsigned long length;
 		file.read((char*)&length, sizeof(length));
